02-add-two-numbers, first successful submission
Problem is to read an integer in reverse from a linked list, and convert
an integer to a linked list in reverse.
Problem involves removing last digit of integer, rounding errors become
problem when that number becomes sufficently large. Convert to string
and back to perform task instead.
Submission statistics:
    faster than: [11%, 17%]
    memory usage: [45%, 99%]
A faster solution?