binary-tree node
Helper for leetcode problems which use a binary-tree, allows conversion
from list to binary-tree (as an instance of 'TreeNode') and back. Empty
list becomes None. None describes an empty node. Lists that have
non-None children of None parents are invalid. __repr__() gives a string
representation of tree. Order of nodes is not enforced. (As yet unused
on leetcode problem).