add safe `_buildTreeFromNestedValuesList`, cleanup
Created safe implementation with gpt4. Still using unsafe version for
now. According to gpt4, a safe version of `_tree_toNestedNodesList` is
not possible without changing return type.

Ongoing: should be a library crate instead of a binary crate (and we
should be importing it properly when using it).

Continue: more testscases (can you get it to break under miri?)