debug `fill_list_infer_missing()`
Function previously incorrect for `values=[1,None,2,None,0,3]`, since it
relies on the length of the input describing the number of levels in the
tree, an incorrect assumption. Added said input to testcases and
debugged. Function now (should) iterate through all of `values`.

Highlighted was the fact there were not enough testcases for this
function - the presumption being it was probably incorrect, yet this
remained neglected until said input finally appeared in a problem.

Continue: more testcases still (for everything).