Changes from python resources/bstreenode
`fill_list_infer_missing()` skip explicit Nones:
If the input is the length of a complete tree, and there are two `None`s
in the input corresponding to where a missing element should be, skip
past them, allowing for the function to correctly process complete
Trees-as-lists where the `None`s to denote empty nodes are not missing.

add list length checks:
Validate tree-as-lists/tree-levels-as-lists are the length they are
supposed to be (one less than a power of two / a power of two) -
throwing sanity-check panics if not.

See: 6b3c4d0f371657fbe5f599e0240363551384e0b9
See: b671d532bf89286637f6159e47fdeff341103208