645-set-mismatch
Given a list of numbers, `s`, containing 1..n, with one number in that
range missing and one duplicated, return as a list, the duplicate and
the missing number

(Problem of the day), easy problem, counter and sorting solutions, and
sorting, invert-duplicate-indexes and mathematical ans. The sorting
solution turned out to be considerably more fiddly than counter
solution. The invert-indexes and mathematical solutions are good
examples of how sometimes easy problems can have solutions that are the
hardest to come up with.