1457-pseudo-palindromic-btree-paths
How many paths from root to each leaf node could be rearranged into a
palindrome.

Solution uses DFS to traverse the tree, with a counter tracking the
number of times each value is encountered for each unique path to a leaf
node, and ans uses bit manipulation to store whether each value has been
seen an odd or even number of times in an integer. Pseudo-palindromic
paths are those which contain either 0 or 1 instances of an odd numbered
count.