225-implement-stack-using-queues
Implement a stack using two queues.

Cheat solution using a list as a stack, and solution using a deque
member variable and temp deque each time an element is removed to remove
the last element from the previous - this is inefficient (O(n) pop), the
only(?) way to remove last element from a queue is to copy everything
but the last element to a new queue? Trying to use two member variable
deque only confused matters - the temp variable makes it much clear what
is going on.

Ans has O(1) pop, it uses nested queues to eliminate need to shuffle
entire contents of the queue every time we remove an element.