1043-partition-array-for-max-sum
Partition input `arr` into contiguous subarrays of at most k elements
such that the sum of the largest element in each subarray multiplied by
that subarray length is maximised.

(Problem of the day), medium problem, implementing bottom-up DP solution
is easy ... once the problem hints give us the definition of the table
and the formula to fill it:
table[i] = max(table[i-j] + max(A[i-1]...A[i-j]) * j) for j=1..=k
One really shouldn't be so quick to resort to hints given in some cases
the essentially contain the answer?
Solution is optimised by storing the max value seen so far for the inner
loop j instead of calling `max(arr[i-j:i])` for each iteration of the
inner loop. Top-down DP solution not attempted - problem seems like a
far better fit for bottom-up DP?