76-min-window-substring
Determine the minimum window substring of `s` containing every character
of `t`.

(Problem of the day), hard problem, sliding window solution, difficulty
was keeping counter correctly synchronised with moving window (increment
counter before incrementing `r` and decrement counter after incrementing
`l`). Once again we were too quick to check the problem hints, which
practically spell out the solution. Solution is slow, since we are
iterating through the counter for every iteration of the sliding window.

Ans uses the same approach, but keeps a counter of how many items the
second counter are >= those in the first to avoid the need to iterate
through the counter.