451-sort-chars-by-frequency
Sort the characters in the input string in decreasing order based on the
frequency of occurrences of that character.

(Problem of the day), medium problem. Two solutions, first uses a bucket
array to count characters, second uses `Counter`, then both use heapsort
on the counts before generating the resulting string. Both solutions are
obvious when one knows both bucket arrays / heapq - the trickiest part
being generating the char2index map for the former. Solution using
counter turns out to be faster in practice.