279-perfect-squares
Find the minimum number of perfect squares that sum to a given value.

(Problem of the day), medium problem, top-down DP solution, three ans:
bottom-up DP, greedy-DFS, and mathematical witchcraft (three square theorem).
Greedy-DFS ans is several orders of magnitude faster than either DP approach,
since it asks the question 'does a solution exist for a given value and count'
instead of 'what is the minimum possible count for this value' it is able to
return the moment it finds any solution instead of having to continue searching
for a smaller count.