276-paint-fence
We are painting a fence of `n` posts `k` different colors. There can be
at most 3 adjacent posts of the same color. How many combinations are
possible?

Medium problem, two ans, bottom-up and top-down DP. Figured out every
detail of bottom-up solution (the definition of `table[i][j]`, the fact
we need to multiply previous values by k-1) except the fact we need to
use the previous two values. The symmetrical nature of `table` makes the
optimisation that it only needs to be a 1d table obvious. Top-down
implementation follows simply from formula of bottom-up ans.

An example of again being too quick to look at ans? Only a single
missing detail was actually needed - as if the formula to fill the table
isn't the entirety of the answer.