294-flip-game-ii
Players take turns replace a single instance of '++' with '--' in the
input. The loser is the first player not to be able to make a
substitution. Determine whether the starting player can guarantee a win.

Recursive-memoize backtracking solution (asks at each stage whether
there is any move the current player can make from which it is no longer
possible for the next player to win), and advanced game-theory
(Sprague-Grundy theorem) ans.