1481-least-unique-after-k-removals
Determine the minimum number of unique elements that can be left after
removing exactly `k elements from input.

(Problem of the day), medium problem, solution sorts the values of a
Counter, subtracting the smallest count at each stage until k<0, a very
simple solution for medium problem. Ans uses the same approach, but uses
heapq to do the sorting - should be faster in theory for very large
inputs but small `k` values, but in practice the overhead of a heap
outweighs this for the submission input.