1642-furthest-building-you-can-reach
Start at building 0, and move from building to building. If the next
building is shorter than the current, use either 1 ladder or a number of
bricks equal to the difference in height. What is the furthest we can
travel?

(Problem of the day), medium problem, Two heapq based solutions, problem
requires us to continue until `bricks` are exhausted, at each stage,
identifying the `ladders` largest buildings encountered so far, first
keeps a heap as large as all the buildings passed so far and is TLE,
second makes the optimisation of allowing the heap to grow no larger
than `ladders` and beats 34%.

An example of not looking at ans, simply working at problem until we
arrive first at the general approach (with only the hint this is
heap/priority-queue problem), then continuing to optimise until finally
we pass the time limit.

Continue: explore problem ans.