128-longest-consecutive-sequence
Find the longest consecutive sequence in an unsorted list in O(n).

From top-interview-150, medium problem, TLE solution, disjointed-set
optimal solution, and ans. For each number we encounter in a sequence,
we track what the smallest number that sequence has ever started with,
and skip any such numbers if we encounter them again. Almost gave up on
disjointed-set solution passing, until realising we could iterate over
`nums_set` instead of `nums`, which made the difference between passing
and not.

This solution appears better than both the official ans and the top ans
from the discussion page (which don't appear to be tracking whether
we've seen a number in a sequence before - which I presume is what the
problem means by 'disjointed-set')?