286-missing-number
Given a list containing n distinct numbers in the range [0,n], return
the only number in the range that is missing.

(Problem of the day), easy problem, Three solutions: sorting,
sweep-and-mark, and mathematical, the second two meeting the requirement
of O(n) time and the last also meeting the requirement of O(1) space.
Three answers: a novel bitwise approach, a set approach which is an
improvement on our sweep approach, and a mathematical approach which
improves on our mathematical solution with Gauss' formula:
sum(range(n+1))=(n*(n+1))//2.

An example of a problem which can be solved through many approaches,
some of which are more advanced than the rating of 'easy' implies.