201-bitwise-and-number-range
Return the bitwise and of all numbers in the range [left,right].

(Problem of the day), medium problem, revisited. Initial attempt to
reduce left/right to much closer values start/end was incorrect.
Solution finds the binary 'common prefix' of the two numbers - the
implementation is ours (and easy once one knows what one needs to find),
but realising that the 'common prefix' is what we need to find required
scanning the top of the answer. Additional answer uses 'Brian
Kernighan's algorithm'.

Takeaway is that some problems should be studied on paper.