1609-even-odd-tree
A binary tree is even-odd if at even even-indexed level, all values are
odd and increasing left to right, and at every odd level, values are
even and decreasing left to right (root is level 0).

(Problem of the day), medium problem, BFS solution adds each node along
with its level and position in that level to the search queue. Optimised
BFS answer uses a nested loop, iterating over an entire level at once in
the inner loop eliminating the need to add anything other than nodes
themselves to the queue.