19-remove-nth-node-from-end-of-list
Remove the n-th node from the end of a linked list and return its head.

(Problem of the day), medium problem, revisited. Two pass solution, and
two-pointers solution which accomplishes the same thing in a single
pass. Only hint required to figure out single pass solution was the
knowledge it involves two pointers.