1750-length-of-string-after-deleting-similar-ends
Given a string, repeatedly remove continuous sequences of characters
from the start and end for so long as the character of those start/end
sequences match, then return the length of the remaining string.

(Weekly premium problem), medium problem, three solutions each using
two-pointers, the first generating a new substring for each pair of
sub-sequences removed, the second making a tail-recursive call for each
pair removed, and the last continuing to iterate so long as pairs
remain. The substring version is the most clear, but generating that
many substrings quickly becomes inefficient. Tail-recursive and
iterative solutions are optimisations derived from substring solution.