206-reverse-linked-list
Reverse a linked-list, returning the new head.

(Problem of the day), easy problem, revisited. Trivial (non in-place)
list solution, and in-place two-pointers, and recursive solutions.

We didn't have the in-place two-pointers solution memorized (should we,
given this is *the* quintessential interview question?), but it wasn't
hard to reason out (first guess was wrong, but fixing it was mostly a
matter of looking at a linked list for a few moments). The fact there is
a recursive version was taken from previous attempt (but nothing else) -
while its implementation is trivially adapted from the iterative
solution.