1614-max-nesting-depth-of-parenthesis
Determine the maximum nested depth of parenthesis in a valid
mathematical expression.

(Problem of the day), easy problem, trivial stack and counter solutions.

Comments call this a 'reading comprehension problem' - making sense of
the elaborate description of the problem took considerably longer than
solving it (a lot of words of 'the expression is a valid mathematical
expression) (which is what makes the solution so trivial).