1544-make-string-great
A good string has no two adjacent letters which are the same character
but different case. Make the string good by removing pairs of adjacent
characters that make the string bad until the string is good.

(Problem of the day), easy problem, two variations of an iterative
solution, and stack based answer.

Our iterative solution proved incorrect when performing only a single
pass, but becomes correct if we only make a single change at a time,
iterating over the modified string at each stage - and constantly having
to generate new strings is slow, whether we store them as a string or
list. But, it used the same general intuition as the stack based answer
(but is incorrect because of the way we track the last letter not
removed?).

Continue: review two-pointers answer.