950-reveal-cards-in-increasing-order
Given a list of numbers, return those numbers in an order such that we
will remove the numbers in sorted order when we repeat the following
steps: 1) remove the next number, 2) move the next number to the end.

(Problem of the day), medium problem, indices queue answer.

Our approach of dividing the list in half and taking from alternating
ends from the second half was incorrect. Solution is simple but not
obvious - this is a problem we should have studied more on-paper.