752-open-the-lock
We start with the combination '0000' on a rotary dial. At each step, we
may rotate one wheel, provided we do not set the dial to one of the
combinations in `deadends`. What is the minimum number of steps to reach
`target`.

(Problem of the day), medium problem, two BFS solutions, first stores
combinations as strings, second stores combinations as integers, and BFS
answer.

Annoyingly, for all the trouble implementing `rotate()` for integer
combinations (included is test function used to write it), storing each
combination as an integer provides an improvement in memory used but not
a significant runtime improvement. Also annoyingly, the BFS based answer
is marginally faster than our solutions.