305-verify-preorder-sequence-bst
Given a list that represents a binary search tree, determine whether the
elements output by a preorder traversal of the tree are in order,
ideally using constant space.

Continued, medium problem, stack, constant-space and recursive answers.

Found both what input list is describing and problem is asking
unintuitive, this is our resorting to the answers after failing to solve
it last time.