133-clone-graph
Given a graph, represented using the class `Node`, created and return a
deep copy.

From top-interview-150, medium problem, BFS solution, simple if
convoluted, create adjacency lists as a dictionary and use it to connect
nodes - competitive on runtime but not on memory usage.

Wrote helper functions to create graph nodes from adjacency lists, and
perform BFS traversal to validate the equality of two graphs - which is
us effectively solving the problem twice.