861-score-after-flipping-matrix
Given an m x n binary matrix, determine the highest possible score after
making any number of moves, a move consists of flipping the bits in a
given row or column, and the score is the sum of each row when expressed
as a binary number.

(Problem of the day), medium problem, naive greedy solution and
optimised greedy answers. Naive greedy solution picks the best row or
column to flip at each stage until there are no more best options to
flip. Answer optimises the first (most significant) column first,
then subsequent columns, the first being long-winded and the second very
succinct.

Answers are radically more efficient than our naive greedy solution,
which is so slow really in the spirit of the question it ought not to
pass.