945-min-increment-to-make-array-unique
Determine the number of times it is necessary to increment an element in
the array to make each element in the array unique.

(Problem of the day), medium problem.

Solution 1 uses a naive approach using a counter and is TLE.

Solution 2 is needlessly complex based on the faulty assumption that a
greedy approach should always update the largest remaining duplicate
value, an iterative approach that iterates over the array in reverse and
for each duplicate, searches for the smallest value it can be replaces
with in the array traversed so far. At each stage, we are left with an
otherwise array with one element in the wrong in the wrong position and
we apply a sort provided by gpt4 for this problem.

Solution 3 uses the hint that we can iterate over a sorted array
forwards, and at each stage, increment the current duplicate by 1. This
approach required the hint of the first sentence of the answer, the
implementation is ours.

Answer uses a counter-array, adding the count-1 for the current element
to the result and adding that to the count for the next element.

Takeaway is this is another problem that would have been better studied
on paper.