1823-find-winner-circular-game
Given a circle of players, labeled [1,n], progress clockwise around the
circle k-1 times at each stage, removing the player landed on and
progressing to the next. Return the last remaining player when starting
at player 1.

(Problem of the day), medium problem. Solution uses a dict to describe
the circle while making removal of elements simple, iterating k steps
around each time an element is to be removed giving us an O(n^2)
solution.

Four answers. First list answer uses the modulus of remaining length to
determine the next index without having to iterate k times to reach it,
then removes that element from the middle of the list at each stage,
simplifying the problem of keeping track of indexes and giving a nicer
algorithm that is O(n).

Second answer uses a queue, moving the current element to the end at
each stage. This is an O(n^2) approach but it is still neater than our
map algorithm.

Third and fourth answers implement an O(n) mathematical solution to the
problem (which is known as the Josephus problem) recursively and
iteratively respectively.

Iterating around a circle while removing elements is a somewhat fiddly
problem (hence our choice of a dictionary to model it), however both
list and queue answers show this fiddly problem has a simpler solution.