2028-find-missing-observations
Find an array of `n` elements between [1,6] which when combined with the
array `rolls` give an average value of `mean`.

Problem of the day, medium problem. Two versions of mathematical
solution - determine value to achieve required average, floor it, fill
result array with this value, and increment values as required if said
value was not an integer. First version each value by 1 until this is
achieved, second increases each value by the largest possible amount to
decrease number of required iterations. Answer is the same as our first
approach.

Mathematical approach easily found on paper, solution was obvious from
there, and from there we optimise the first solution to create the
second.