884-uncommon-words-in-two-sentences
Given two sentences, determine the words which occur exactly once in one
sentence and do not occur in the other.

Problem of the day, easy problem. Two solutions and answer. All
approaches use Counter and set operations - first solution generating
extra sets for single and multiple occurrence words for each sentence,
while the second omits this step simplifying the logic (and is no slower
for doing so). The answer provides the insight that we only need a
single counter for both sentences combined, since the result will be any
words with a count of 1 in this combined counter.