2684-max-number-moves-in-grid
Determine the maximum number of times we can move left-right through a
grid if for each move we must move onto a larger square.

Problem of the day, medium problem. Two dynamic-programming solutions,
bottom-up and top-down. Bottom-up table solution is relatively
straightforward, and top-down recursive-memoization solution is simply
adapted from that. The bottom-up solution is relatively slow but memory
efficient, the memoization solution is the opposite.

However, the optimum solution is a BFS which we see consulting answers.
Continue: review BFS answer