review BFS answer
BFS answer relies on the fact that the longest route to a given column
is the index of that column because we may only move left-right. We
perform a BFS through the grid until we reach the RHS, or run out of
squares.

Cleaned up previous solutions by combining up-straight-down cases into a
single delta-direction loop.