2461-max-sum-distinct-subarrays-len-k
Find the contiguous subarray of unique values of length `k` in `nums`
with the highest sum.

Problem of the day, medium problem. Sliding window with running counter
and running sum solution (passes once we track the window sum and the
window separately, calculating the sum each time is O(n*k) and times
out). Not an optimal solution only beating 11%.

Continue: review num-to-index-map answer.